HW-3 패키지 구조 재수정
