+++ 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/login/LoginRepository.kt
@@ -1,0 +1,4 @@
+package com.mtjin.androidarchitecturestudy.data.login
+
+interface LoginRepository {
+}
