+++ 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/login/source/local/LoginLocalDataSourceImpl.kt
@@ -1,0 +1,4 @@
+package com.mtjin.androidarchitecturestudy.data.login.source.local
+
+class LoginLocalDataSourceImpl {
+}
