+++ 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/movie_search/MovieResponse.kt
@@ -1,7 +1,8 @@
-package com.mtjin.androidarchitecturestudy.data
+package com.mtjin.androidarchitecturestudy.data.movie_search
 
 
 import com.google.gson.annotations.SerializedName
+import com.mtjin.androidarchitecturestudy.data.movie_search.Movie
 
 data class MovieResponse(
     @SerializedName("display")
