+++ 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/data/movie_search/source/local/MovieDao.kt
@@ -1,10 +1,10 @@
-package com.mtjin.androidarchitecturestudy.data.source.local.movie_search
+package com.mtjin.androidarchitecturestudy.data.movie_search.source.local
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.mtjin.androidarchitecturestudy.data.Movie
+import com.mtjin.androidarchitecturestudy.data.movie_search.Movie
 
 @Dao
 interface MovieDao {
