+++ 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/ui/search/MovieSearchPresenter.kt
@@ -1,7 +1,7 @@
 package com.mtjin.androidarchitecturestudy.ui.search
 
 import android.util.Log
-import com.mtjin.androidarchitecturestudy.data.source.MovieRepository
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.MovieRepository
 import retrofit2.HttpException
 
 class MovieSearchPresenter(
