+++ 2003/mtjin/AndroidArchitectureStudy/app/src/main/java/com/mtjin/androidarchitecturestudy/utils/MyApplication.kt
@@ -3,14 +3,14 @@ package com.mtjin.androidarchitecturestudy.utils
 import android.app.Application
 import com.mtjin.androidarchitecturestudy.api.ApiClient
 import com.mtjin.androidarchitecturestudy.api.ApiInterface
-import com.mtjin.androidarchitecturestudy.data.source.MovieRepository
-import com.mtjin.androidarchitecturestudy.data.source.MovieRepositoryImpl
-import com.mtjin.androidarchitecturestudy.data.source.local.movie_search.MovieDao
-import com.mtjin.androidarchitecturestudy.data.source.local.movie_search.MovieDatabase
-import com.mtjin.androidarchitecturestudy.data.source.local.movie_search.MovieLocalDataSource
-import com.mtjin.androidarchitecturestudy.data.source.local.movie_search.MovieLocalDataSourceImpl
-import com.mtjin.androidarchitecturestudy.data.source.remote.movie_search.MovieRemoteDataSource
-import com.mtjin.androidarchitecturestudy.data.source.remote.movie_search.MovieRemoteDataSourceImpl
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.MovieRepository
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.MovieRepositoryImpl
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.local.MovieDao
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.local.MovieDatabase
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.local.MovieLocalDataSource
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.local.MovieLocalDataSourceImpl
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.remote.MovieRemoteDataSource
+import com.mtjin.androidarchitecturestudy.data.movie_search.source.remote.MovieRemoteDataSourceImpl
 
 class MyApplication : Application() {
     private lateinit var networkManager: NetworkManager
@@ -33,10 +33,7 @@ class MyApplication : Application() {
             MovieRemoteDataSourceImpl(
                 apiInterface
             )
-        movieLocalDataSource =
-            MovieLocalDataSourceImpl(
-                movieDao
-            )
+        movieLocalDataSource = MovieLocalDataSourceImpl(movieDao)
         movieRepository = MovieRepositoryImpl(movieRemoteDataSource, movieLocalDataSource, networkManager)
     }
 }
